# Speed up dlpack export and ignore stream for CPU requests